Aggiornato file per uso della stampa con gnuplot e aggiunti comandi gnuplot
