+++ PRO2/PRO2.vcxproj
@@ -21,6 +21,7 @@
   <ItemGroup>
     <ClCompile Include="comandiGnuplot.c" />
     <ClCompile Include="Main.c" />
+    <ClCompile Include="Parser.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="TSP.h" />
@@ -91,6 +92,7 @@
       <Optimization>Disabled</Optimization>
       <SDLCheck>true</SDLCheck>
       <ConformanceMode>true</ConformanceMode>
+      <PreprocessorDefinitions>_MBCS;%(PreprocessorDefinitions) _CRT_SECURE_NO_WARNINGS</PreprocessorDefinitions>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
