+++ PRO2/PRO2.vcxproj.filters
@@ -13,7 +13,7 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
-    <Filter Include="FunctionGnuplot">
+    <Filter Include="FileFunzioni">
       <UniqueIdentifier>{8d8c91bc-86d2-43d5-9511-1254ad54e3b0}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
@@ -22,7 +22,10 @@
       <Filter>File di origine</Filter>
     </ClCompile>
     <ClCompile Include="comandiGnuplot.c">
-      <Filter>FunctionGnuplot</Filter>
+      <Filter>FileFunzioni</Filter>
+    </ClCompile>
+    <ClCompile Include="Parser.c">
+      <Filter>FileFunzioni</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
